- Move all the rules into `Rule` a directory. - Delete the unused `Data.Queue`.
