+++ specs/chain/hs/src/Cardano/Spec/Chain/STS/Rule/Epoch.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE TypeFamilies #-}
 
-module Cardano.Spec.Chain.STS.Epoch where
+module Cardano.Spec.Chain.STS.Rule.Epoch where
 
 import Data.Map.Strict (Map)
